+++ src/libs/index.js
@@ -1,0 +1,5 @@
+export http from './http';
+export httpProvider from './http-provider';
+export Stream from './stream';
+export StreamManager from './stream-manager';
+export StreamComponent from './stream-component';
