+++ src/libs/stream-component.js
