+++ src/libs/stream-manager.js
