+++ src/libs/stream.js
